backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-dotenv==1.0.0

librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

google-cloud-speech==2.24.0
openai==1.12.0
transformers==4.36.0
torch==2.1.0

deepgram-sdk==3.0.0
assemblyai==0.17.0

python-docx==1.1.0
reportlab==4.0.9

requests==2.31.0

playwright==1.41.0
pyvirtualdisplay==3.0
